Extract JWT to it's own file
